[ui][events] Use explicitly-sized type for KeyboardCode enums
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7653076
Auto-Submit: Wez <wez@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1597732}

+++ ui/events/keycodes/keyboard_codes_win.h
@@ -8,11 +8,13 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef UI_EVENTS_KEYCODES_KEYBOARD_CODES_WIN_H_
 #define UI_EVENTS_KEYCODES_KEYBOARD_CODES_WIN_H_
 
+#include <cstdint>
+
 #include "ui/events/keycodes/keyboard_defines_win.h"
 
 namespace ui {
 
-enum KeyboardCode : unsigned short {
+enum KeyboardCode : uint8_t {
   VKEY_CANCEL = VK_CANCEL,
   VKEY_BACK = VK_BACK,
   VKEY_TAB = VK_TAB,
